backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
pyautogen
python-dotenv
pydantic
